requirements.txt
certifi==2019.3.9
chardet==3.0.4
dj-database-url==0.5.0
django-bootstrap3==11.0.0
django-bootstrap3-datepicker==0.4.1
django-ckeditor==5.6.1
django-cors-headers==2.5.0
django-crispy-forms==1.6.0
django-datetimepicker==3.14
django-debug-toolbar==1.11
django-heroku==0.3.1
django-js-asset==1.2.2
django-markdown-deux==1.0.5
django-multiselectfield==0.1.8
django-pagedown==0.1.1
djangorestframework==3.9.2
gunicorn==19.9.0
idna==2.6
markdown2==2.3.7
mysqlclient==1.4.6
Naked==0.1.31
Pillow==5.4.0
psycopg2==2.7.7
PyMySQL==0.9.3
python-nmap==0.6.1
pytz==2018.7
PyYAML==3.12
requests==2.18.4
shellescape==3.4.1
sqlparse==0.2.4
urllib3==1.22
virtualenv==16.0.0
virtualenvwrapper-win==1.2.5
Werkzeug==0.14.1
whitenoise==4.1.2
social-auth-app-django==3.1.0
